Move tests.test_legacy.DocTestCase to tests.mongo.test_documents.DocumentBaseTestCase
